requirements.txt
streamlit==1.25.0
pandas==2.1.1
requests==2.31.0
mysql-connector-python==8.1.0
pytest==7.4.0
coverage==7.3.2
alembic==1.11.1
SQLAlchemy==2.0.21